Move completedBacklogs page logic to viewmodel
